requirements.txt
django-tinymce
django-weasyprint
